+++ src/main/java/org/jabref/model/database/BibDatabaseContext.java
@@ -10,6 +10,8 @@ import java.util.stream.Collectors;
 
 import org.jabref.architecture.AllowedToUseLogic;
 import org.jabref.gui.LibraryTab;
+import org.jabref.logic.crawler.Crawler;
+import org.jabref.logic.crawler.StudyRepository;
 import org.jabref.logic.shared.DatabaseLocation;
 import org.jabref.logic.shared.DatabaseSynchronizer;
 import org.jabref.logic.util.CoarseChangeFilter;
@@ -17,6 +19,7 @@ import org.jabref.logic.util.OS;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.metadata.MetaData;
 import org.jabref.model.pdf.search.SearchFieldConstants;
+import org.jabref.model.study.Study;
 import org.jabref.preferences.FilePreferences;
 
 import net.harawata.appdirs.AppDirsFactory;
@@ -111,6 +114,16 @@ public class BibDatabaseContext {
 
     public boolean isBiblatexMode() {
         return getMode() == BibDatabaseMode.BIBLATEX;
+    }
+
+    /**
+     * Returns whether this .bib file belongs to a {@link Study}
+     */
+    public boolean isStudy() {
+        return this.getDatabasePath()
+                .map(path -> path.getFileName().toString().equals(Crawler.FILENAME_STUDY_RESULT_BIB) &&
+                        Files.exists(path.resolveSibling(StudyRepository.STUDY_DEFINITION_FILE_NAME)))
+                .orElse(false);
     }
 
     /**